🔨 rename endpoint types folder
