+++ packages/@ourworldindata/types/src/index.ts
@@ -822,4 +822,4 @@ export {
     type GrapherValuesJson,
     type GrapherValuesJsonDataPoints,
     type GrapherValuesJsonDataPoint,
-} from "./functionTypes/GrapherEndpoints.js"
+} from "./endpointTypes/GrapherEndpoints.js"
